requirements.manual.txt
django
djangorestframework
django-cors-headers
django-filter
python-dotenv
djangorestframework-simplejwt
